Q9 benchmark

+++ src/test/scala/lms/tutorial/dslapi.scala
@@ -222,9 +222,9 @@ trait DslGenC extends CGenNumericOps
       #ifndef MAP_FILE
       #define MAP_FILE MAP_SHARED
       #endif
-      int fsize(int fd) {
+      int64_t fsize(int fd) {
         struct stat stat;
-        int res = fstat(fd,&stat);
+        int64_t res = fstat(fd,&stat);
         return stat.st_size;
       }
       int printll(char* s) {
